tighter calc

+++ src/main/java/datawave/data/normalizer/AbstractGeometryNormalizer.java
@@ -216,10 +216,10 @@ public abstract class AbstractGeometryNormalizer<T extends datawave.data.type.ut
     public boolean validLength(short tier, String value) {
         // determine the length of the position in hex characters
         // ceil(tier/4) will get the number of bytes
-        long bytes = (tier >> 2) + ((tier & 0x3) == 0 ? 0 : 1);
+        int bytes = (tier >> 2) + ((tier & 0x3) == 0 ? 0 : 1);
         
         // multiply by 2 to get the number of hex digits
-        long posLen = 2L * bytes;
+        int posLen = 2 * bytes;
         // length is the tier length plus the position length
         return value.length() == (2 + posLen);
     }
